pyproject.toml
[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.py$'
force-exclude = '''
/(
    | rare/ui
    | rare/legendary
    | rare/resources
)/
'''

[tool.poetry]
name = "rare"
version = "1.10.11"
description = "A gui for legendary"
authors = ["RareDevs"]
license = "GPL3"
readme = "README.md"
repository = "https://github.com/RareDevs/Rare"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
PySide6-Essentials = "^6.5.3"
QtAwesome = "^1.1.1"
legendary-gl = "^0.20.34"
pywebview = [
    { version = "^3.6.3", extras = ["cef"], platform = "windows", optional = true },
    { version = "^3.6.3", extras = ["gtk"], platform = "linux", optional = true },
    { version = "^3.6.3", extras = ["gtk"], platform = "freebsd", optional = true },
]
orjson = "^3.8.0"
vdf = "^3.4"
pywin32 = { version = "^306", markers = "platform_system == 'Windows'" }
pypresence = { version = "^4.2.1", optional = true }

[tool.poetry.scripts]
start = "rare.main:main"

[tool.poetry.dev-dependencies]
Nuitka = "^2.4.8"
pylint = "^3.2.7"
black = "^24.3.0"
PySide6-stubs = "^6.4.2.0"

#[build-system]
#requires = ["setuptools>=42", "wheel", "nuitka", "toml"]
#build-backend = "nuitka.distutils.Build"

[nuitka]
assume-yes-for-downloads = true
mingw64 = true
lto = false
jobs = 4
static-libpython = false
standalone = true
show-scons = false
enable-plugin = ["anti-bloat", "pyside6"]
show-anti-bloat-changes = true
follow-stdlib = true
follow-imports = true
nofollow-import-to = ["*.tests", "*.distutils", "distutils", "unittest", "pydoc", "tkinter", "test"]
prefer-source-code = true
include-package = "pypresence"
include-package-data = "qtawesome"
include-data-dir = "rare/resources/images/=rare/resources/images/"
include-data-files = "rare/resources/languages/=rare/resources/laguanges/=rare_*.qm"
windows-icon-from-ico = "rare/resources/images/Rare.ico"
windows-company-name = "Rare"
windows-product-name = "Rare"
windows-file-version = "1.10.11.0"
windows-product-version = "1.10.11.0"
windows-disable-console = true
